+++ R/adae.R
@@ -52,8 +52,6 @@
 #'  \item{ADURN}{Analysis Duration (N)}
 #'  \item{ADURU}{Analysis Duration Units}
 #'  \item{LDOSEDTM}{End Date/Time of Last Dose}
-#'  \item{DOSEON}{Treatment Dose at Record Start}
-#'  \item{DOSEU}{Treatment Dose Units}
 #'  \item{ASEV}{Analysis Severity/Intensity}
 #'  \item{AREL}{Analysis Causality}
 #'  \item{TRTEMFL}{Treatment Emergent Analysis Flag}
@@ -63,6 +61,8 @@
 #'  \item{AETOXGRN}{Standard Toxicity Grade (N)}
 #'  \item{AEACN_DECODE}{Action Taken with Study Treatment}
 #'  \item{DOSEDY}{Day of Study Drug}
+#'  \item{DOSEU}{Treatment Dose Units}
+#'  \item{DOSEON}{Treatment Dose at Record Start}
 #'  \item{AECONTRT}{Concomitant or Additional Trtmnt Given}
 #'  \item{CQ01NAM}{Customized Query 01 Name}
 #'  \item{CQ02NAM}{Customized Query 02 Name}
@@ -87,7 +87,7 @@
 #'  \item{TRT01P}{Planned Treatment for Period 01}
 #'  \item{TRDISCFL}{Treatment Discontinued Flag}
 #' }
-#' @seealso \code{\link{adae}} \code{\link{adaeocmq}} \code{\link{adcm}} \code{\link{adeg}} \code{\link{adex}} \code{\link{adexsum}} \code{\link{adlb}} \code{\link{adpc}} \code{\link{adsl}} \code{\link{adttesaf}} \code{\link{advs}}# nolint
+#' @seealso \code{\link{adae}} \code{\link{adaeocmq}} \code{\link{adagocmq}} \code{\link{adcm}} \code{\link{adeg}} \code{\link{adex}} \code{\link{adexsum}} \code{\link{adlb}} \code{\link{adpc}} \code{\link{adsl}} \code{\link{adttesaf}} \code{\link{advs}}# nolint
 #' @keywords datasets adae
 #' @name adae
 #' @examples