+++ R/adex.R
@@ -79,15 +79,15 @@
 #'  \item{SCRFDT}{Screen Failure Date}
 #'  \item{EOSDT}{End of Study Date}
 #'  \item{EOSSTT}{End of Study Status}
-#'  \item{FRVDT}{Final Retrieval Visit Date}
+#'  \item{FRVDT}{Final Retrievel Visit Date}
 #'  \item{RANDDT}{Date of Randomization}
 #'  \item{DTHDT}{Date of Death}
-#'  \item{DTHDTF}{Date of Death Imputation Flag}
+#'  \item{DTHDTF}{DTHDTF}
 #'  \item{DTHADY}{Relative Day of Death}
 #'  \item{LDDTHELD}{Elapsed Days from Last Dose to Death}
-#'  \item{DTHCAUS}{Cause of Death}
-#'  \item{DTHDOM}{Domain for Date of Death Collection}
-#'  \item{DTHCGR1}{Cause of Death Reason 1}
+#'  \item{DTHCAUS}{DTHCAUS}
+#'  \item{DTHDOM}{DTHDOM}
+#'  \item{DTHCGR1}{DTHCGR1}
 #'  \item{LSTALVDT}{Date Last Known Alive}
 #'  \item{SAFFL}{Safety Population Flag}
 #'  \item{RACEGR1}{Pooled Race Group 1}
@@ -134,7 +134,7 @@
 #'  \item{AINFRAT}{Analysis Infusion Rate}
 #'  \item{AINFRAU}{Analysis Infusion Rate Unit}
 #' }
-#' @seealso \code{\link{adae}} \code{\link{adaeocmq}} \code{\link{adcm}} \code{\link{adeg}} \code{\link{adex}} \code{\link{adexsum}} \code{\link{adlb}} \code{\link{adpc}} \code{\link{adsl}} \code{\link{adttesaf}} \code{\link{advs}}# nolint
+#' @seealso \code{\link{adae}} \code{\link{adaeocmq}} \code{\link{adagocmq}} \code{\link{adcm}} \code{\link{adeg}} \code{\link{adex}} \code{\link{adexsum}} \code{\link{adlb}} \code{\link{adpc}} \code{\link{adsl}} \code{\link{adttesaf}} \code{\link{advs}}# nolint
 #' @keywords datasets adex
 #' @name adex
 #' @examples